requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-redis
dj-database-url
drf-yasg